chatbot/requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.25.0
pydantic==2.5.0
qdrant-client==1.8.0
transformers==4.36.0
torch==2.1.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-asyncio==0.23.0
httpx==0.25.0
numpy==1.24.0